src/fortiusant/requirementsNoGUI.txt
bless
lib_programname
numpy != 1.19.4
pyinstaller
pyusb
requests
service_identity
wheel
dbus_next
